requirements.txt
opencv-python
opencv-python-headless
ultralytics
pyserial
requests
datetime
